requirements.txt
PyYAML==5.3.1
aws-cdk.core
